@@ -1,5 +1,5 @@
 {
-  "name": "portfolio",
+  "name": "raymarching",
   "version": "0.0.0",
   "private": true,
   "scripts": {
@@ -15,22 +15,27 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.14",
-    "@docusaurus/preset-classic": "2.0.0-beta.14",
-    "@mdx-js/react": "^1.6.21",
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
+    "@mdx-js/react": "^3.0.0",
     "clsx": "^1.1.1",
     "hast-util-is-element": "^1.1.0",
-    "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "rehype-katex": "^4.0.0",
-    "remark-math": "^3.0.1",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-player": "^2.16.0",
+    "rehype-katex": "^7.0.0",
+    "remark-math": "^6.0.0",
     "twgl.js": "^4.22.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "2.0.0-beta.14",
+    "@docusaurus/module-type-aliases": "3.0.0",
+    "@docusaurus/types": "3.0.0",
     "@tsconfig/docusaurus": "^1.0.4",
     "typescript": "^4.5.2"
+  },
+  "engines": {
+    "node": ">=18.0"
   },
   "browserslist": {
     "production": [
